ENH: Print message if error occurs when deleting the job in the execution server
The job can be deleted only in the server side when we are sure that the data has been removed from the execution server

+++ src/clusterpost-provider/dataprovider.handlers.js
@@ -196,8 +196,15 @@ module.exports = function (server, conf) {
 		})
 		.then(function(doc){
 			return server.methods.executionserver.jobdelete(doc)
-			.then(function(){
-				return server.methods.clusterprovider.deleteDocument(doc);
+			.then(function(res){
+				if(res.indexOf("Deleted:" + doc._id)){
+					return server.methods.clusterprovider.deleteDocument(doc);
+				}else{
+					return {
+						error: "Job data in execution server could not be deleted.",
+						status: res
+					};
+				}
 			});
 		})
 		.then(rep)
